adbtool: split into modules
